refactor: move useFoxyQuery
