+++ src/features/defi/providers/foxy/components/FoxyManager/Deposit/components/Approve.tsx
@@ -7,7 +7,7 @@ import type { DepositValues } from 'features/defi/components/Deposit/Deposit'
 import { DefiAction, DefiStep } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { useFoxy } from 'features/defi/contexts/FoxyProvider/FoxyProvider'
 import { canCoverTxFees } from 'features/defi/helpers/utils'
-import { useFoxyQuery } from 'features/defi/providers/foxy/components/FoxyManager/Deposit/components/useFoxyQuery'
+import { useFoxyQuery } from 'features/defi/providers/foxy/components/FoxyManager/useFoxyQuery'
 import { useCallback, useContext, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { useHistory } from 'react-router-dom'
