+++ src/features/defi/providers/foxy/components/FoxyManager/useFoxyQuery.ts
